Improve drive power clipping

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Constants.java
@@ -1,8 +1,8 @@
 package org.firstinspires.ftc.teamcode;
 
 public class Constants {
-    public static final float DRIVE_MAX = 0.8f;
-    public static final float DRIVE_MIN = -0.8f;
+    public static final double DRIVE_MAX = 0.8;
+    public static final double DRIVE_MIN = -0.8;
 
     public static final double LEFT_FOUNDATION_INIT = 0.0;
     public static final double LEFT_FOUNDATION_MAX = 0.70;
